<App> component move to ./app dir.
